requirements.txt
alembic==0.9.1
appdirs==1.4.3
humanize==0.5.1
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-editor==1.0.3
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.9
